Revised to use character abilities
Substantial re-write of creating the action choice dialog and parsing the user input. Switches allow each action choice to be toggled on or off, and most can be set to repeat any number of times. (So attack, for example, could be displayed 3 times.)

When displaying choices for a single character, pull character attributes where appropriate. 

Include calculation of character attributes, such as dexterity bonus.